+++ OpenAI.Responses/src/Custom/Items/ShellTool/ShellCallItem.cs
@@ -1,0 +1,14 @@
+using Microsoft.TypeSpec.Generator.Customizations;
+
+namespace OpenAI.Responses;
+
+// CUSTOM: Renamed.
+[CodeGenType("ShellToolCallItemResource")]
+public partial class ShellCallItem
+{
+    // CUSTOM:
+    // - Made nullable because this is an optional property.
+    // - Added setter because this is an optional property in an input/output type.
+    [CodeGenMember("Status")]
+    public ShellCallStatus? Status { get; set; }
+}